merge sort archived and chapter 2 beginnings
I gave up on visualising merge sort for the moment. I decided to go two
chapters back and finally implement the positional list ADT. I was able
to add all the buttons for the operations present in the ADT. I was also
able to dynamically add an element to the page using js.